🔒 Fix Uncontrolled Resource Consumption in Pascal's Triangle
Refactored `generate_pascals_triangle` to match the exact requirements of the vulnerability report by renaming the input parameter `n` to `num_rows`. Maintained the existing hard upper bound of 1000 to prevent uncontrolled resource consumption (DoS through O(N^2) arbitrary-precision memory allocation).

Additionally, extracted related tests from the monolithic `test_DiscreteMath.py` into a new, dedicated `test_pascals_triangle.py` according to project standards.
