+++ Tests/test_DiscreteMath.py
@@ -25,7 +25,6 @@ from Math.Discrete_Math.Combinatorics.combination import nCr
 
 from Math.Discrete_Math.Combinatorics.binomial_theorem import expand_binomial, binomial_coefficient
 from Math.Discrete_Math.Combinatorics.binomial_theorem_general_term import binomial_general_term
-from Math.Discrete_Math.Combinatorics.pascals_triangle import print_pascals_triangle, generate_pascals_triangle
 from Math.Discrete_Math.Combinatorics.permutation import factorial, n_permute_r
 from Math.Discrete_Math.Combinatorics.trinomial_theorem import expand_trinomial, trinomial_coefficient
 from Math.Discrete_Math.Combinatorics.trinomial_theorem_general_term import trinomial_general_term
@@ -127,44 +126,6 @@ class TestPermutation(unittest.TestCase):
     def test_n_permute_r_invalid(self):
         with self.assertRaises(ValueError):
             n_permute_r(3, 5)
-
-
-class TestPascalsTriangle:
-    def test_print_pascals_triangle_normal(self, capsys):
-        """Test printing a normal Pascal's triangle."""
-        triangle = [[1], [1, 1], [1, 2, 1]]
-        print_pascals_triangle(triangle)
-        captured = capsys.readouterr()
-
-        expected_output = "  1  \n 1 1 \n1 2 1\n"
-        assert captured.out == expected_output
-
-    def test_print_pascals_triangle_empty(self, capsys):
-        """Test printing an empty Pascal's triangle."""
-        triangle = []
-        print_pascals_triangle(triangle)
-        captured = capsys.readouterr()
-
-        assert captured.out == ""
-
-    def test_generate_pascals_triangle_valid(self):
-        """Test generating a valid Pascal's triangle."""
-        assert generate_pascals_triangle(0) == []
-        assert generate_pascals_triangle(1) == [[1]]
-        assert generate_pascals_triangle(2) == [[1], [1, 1]]
-        assert generate_pascals_triangle(3) == [[1], [1, 1], [1, 2, 1]]
-        assert generate_pascals_triangle(5) == [[1], [1, 1], [1, 2, 1], [1, 3, 3, 1], [1, 4, 6, 4, 1]]
-
-    def test_generate_pascals_triangle_invalid(self):
-        """Test generating an invalid Pascal's triangle."""
-        with pytest.raises(ValueError, match="Number of rows cannot be negative."):
-            generate_pascals_triangle(-1)
-        with pytest.raises(ValueError, match="Number of rows cannot be negative."):
-            generate_pascals_triangle(-5)
-        with pytest.raises(ValueError):
-            generate_pascals_triangle(1001)
-        with pytest.raises(ValueError):
-            generate_pascals_triangle(2000)
 
 
 def test_prime_factorization_edge_case():
